+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/api/GetGameHighScores.kt
@@ -3,6 +3,7 @@
 package eu.vendeli.tgbot.api
 
 import eu.vendeli.tgbot.interfaces.Action
+import eu.vendeli.tgbot.interfaces.InlineActionExt
 import eu.vendeli.tgbot.types.User
 import eu.vendeli.tgbot.types.game.GameHighScore
 import eu.vendeli.tgbot.types.internal.Identifier
@@ -12,7 +13,7 @@ import eu.vendeli.tgbot.utils.getReturnType
 import eu.vendeli.tgbot.utils.serde.DynamicLookupSerializer
 import eu.vendeli.tgbot.utils.toJsonElement
 
-class GetGameHighScoresAction : Action<List<GameHighScore>> {
+class GetGameHighScoresAction : Action<List<GameHighScore>>, InlineActionExt<List<GameHighScore>> {
     override val method = TgMethod("getGameHighScores")
     override val returnType = getReturnType()